Automaton

Documentation

data FSM a

Finite state machine / Automaton, consisting of a label and a set of transitions.

Constructors

FSM

Fields

label :: String

The label, or name of the automaton.

transitions :: Set (T a)

The set of transitions of the automaton. ^ Since they contain the states, we don't include ^ the set of all states in the structure itself, ^ as it can be obtained through the transitions.

fsmAddTrans :: Ord a => FSM a -> Set (T a) -> FSM a

Add a set of transitions to an automaton.

fsmTransFromState :: Eq a => S a -> FSM a -> Set (T a)

Get the transitions having this state as a starting point.

fsmTransFromSetState :: Ord a => S (Set a) -> FSM a -> Set (T a)

Get the transitions having some state as a starting point, from all the states that this state represents.

fsmTransFromPairState :: (Eq a, Eq b) => S (a, b) -> (FSM a, FSM b) -> (Set (T a), Set (T b))

Get the paired transitions from a pair state in a pair of automata.

fsmFromList :: Ord a => String -> [T a] -> FSM a

Create an automaton from a list of transitions.

fsmStates :: Ord a => FSM a -> Set (S a)

Get the states of an automaton.

fsmInitialStates :: Ord a => FSM a -> Set (S a)

Get the initial states of an automaton.

fsmInitialState :: Ord a => FSM a -> Maybe (S a)

Get the supposedly unique initial state of an automaton. Errors out if there are none.

fsmFinalStates :: Ord a => FSM a -> Set (S a)

Get the final states of an automaton.

succState :: Ord a => FSM a -> Char -> S a -> Set (S a)

Get all the succeeding states from a state and a symbol.

allSuccState :: Ord a => FSM a -> String -> S a -> [Set (S a)]

Get all the succeeding states from a state and an alphabet.

succStates :: Ord a => FSM a -> Char -> Set (S a) -> Set (S a)

Get all the succeeding states from a set of states and a symbol.

fsmExec :: Ord a => FSM a -> String -> Bool

Execute an automaton on a series of symbols.

isComplete :: Ord a => String -> FSM a -> Bool

Whether the automaton is complete.

symbolsWithStates :: Ord a => FSM a -> [(S a, String)]

Get all individual symbols for each state from an automaton, with the states.

symbolsFromStates :: Ord a => FSM a -> [String]

Get all individual symbols for each state from an automaton, without the states.

fsmAlphaFromState :: Eq a => S a -> FSM a -> String

Get the alphabet from the state of an automaton.

isDeterministic :: Ord a => FSM a -> Bool

Whether the automaton is deterministic.

complete :: (Ord a, Show a) => a -> String -> FSM a -> FSM a

Outputs a completed version of the automaton.

type NewTransGen a w

Arguments

= S a
The current state to generate transitions for.

-> Set (S a)
The stack of handled states.

-> w
Whatever object is used to generate the transitions from.

-> Set (T a)

New transition generator. Used to generate automata by construction, such as DFA conversion, intersection and union.

generateTransFn :: Ord a => NewTransGen a w -> NewTransGen a w

Generates a recursive transition generator.

generateFSMFn :: Ord a => String -> S a -> NewTransGen a w -> w -> FSM a

Generates an automaton by construction, starting from an initial state using a recursive transition generator.

generateDFATrans :: Ord a => NewTransGen (Set a) (FSM a)

Recursive transition generator for the conversion to DFA.

convertToDFA :: (Show a, Ord a) => FSM a -> FSM (Set a)

Converts an automaton to a DFA.

generatePairTrans

Arguments

:: (Ord a, Ord b)

=> BoolRelation
The pairing operation.

-> NewTransGen (a, b) (FSM a, FSM b)

Recursive transition generator for a pairing operation of two automata.

generatePairFSM

Arguments

:: (Ord a, Ord b)

=> String
The label separator.

-> BoolRelation
The pairing operation, for example (&&) or (||).

-> (FSM a, FSM b)
The pair of two automata.

-> FSM (a, b)

Constructs two automata with a pairing operation using a pair of two automata.

(<&>) :: (Ord a, Ord b) => FSM a -> FSM b -> FSM (a, b)

Outputs the intersection of two automata. # Hypothesis Both have to be deterministic, but not necessarily complete. However, the function doesn't check.

(<|>) :: (Ord a, Ord b) => FSM a -> FSM b -> FSM (a, b)

Outputs the union of two automata. # Hypothesis Both have to be deterministic and complete. However, the function doesn't check.

(<^>) :: (Ord a, Ord b) => FSM a -> FSM b -> FSM (a, b)

Outputs the xor of two automata. # Hypothesis Both have to be deterministic and complete. However, the function doesn't check.

(<=>) :: (Ord a, Ord b) => FSM a -> FSM b -> FSM (a, b)

Outputs the equivalence of two automata. # Hypothesis Both have to be deterministic and complete. However, the function doesn't check.