DotShow

Documentation

class DotShow a where

Typeclass similar to Show used to show the graph representation of automata using DOT.

Minimal complete definition

dotShow

Methods

dotShow :: a -> String

Outputs a string written in the DOT language to show the graph representation of a.

dotDeclare :: a -> Maybe String

Special case where we want to declare some objects like states inside a DOT file.

dotPDF :: a -> IO ()

Generates and opens a PDF showing the graph representation of a (with default filename).

dotShowPDF

Arguments

:: String
The name of the file to generate.

-> a

-> IO ()

Generates and opens a PDF showing the graph representation of a.