Helpers

Documentation

clr :: String

ANSI color code to reset the styles.

bld :: String

ANSI color code to make the text bold.

dim :: String

ANSI color code to make the text dim.

red :: String

ANSI color code to make the text red.

grn :: String

ANSI color code to make the text green.

yel :: String

ANSI color code to make the text yellow.

(<<$>>) :: Ord b => (a -> b) -> Set a -> Set b

Map operator for sets, which can't implement the Functor typeclass due to type constraints.

(<>>=>) :: (Foldable m, Foldable n, Monoid (n b)) => m a -> (a -> n b) -> n b

Flatmap operator for sets, which can't implement the Monad typeclass due to type constraints.

hasDuplicates :: (Eq a, Ord a) => [a] -> Bool

Whether a list has duplicates or not. Constructs a set and compares its size to the original list's size.

alphaComp

Arguments

:: (Foldable t, Eq a)

=> t a
The alphabet.

-> [a]
The other partition.

-> [a]

Get the complementary partition of an alphabet.

Example: alphaComp "abcd" "bd" == "ac"

zipMaybe :: (Maybe a, Maybe b) -> Maybe (a, b)

Zips a pair of maybes into a Maybe of a pair. It will be Just (a, b) when a and b are something, and Nothing otherwise.