State

Documentation

data S a

State of an automaton.

Constructors

S

Fields

stateLabel :: a

Label of the state, which can be of any type. Useful for various operations on automata.

stateFlags :: (Bool, Bool)

Flags of the state (initial, final).

initialStates :: Set (S a) -> Set (S a)

Get the initial states from the set of states.

initialState :: Set (S a) -> Maybe (S a)

Get the unique initial state from the set of states. Returns Nothing if there is no initial state.

isInitialSet :: Set (S a) -> Bool

Checks if a set of states contains initial states.

finalStates :: Set (S a) -> Set (S a)

Get the final states from the set of states.

isFinalSet :: Set (S a) -> Bool

Checks if a set of states contains final states.

mergeStates :: Ord a => Bool -> Set (S a) -> S (Set a)

Merges a set of states into a state of sets. This is where an arbitrary typed label for the set is useful.

type BoolRelation = Bool -> Bool -> Bool

Boolean relation: two booleans being related or not.

pairStates :: Ord a => Bool -> BoolRelation -> (S a, S b) -> S (a, b)

Pairs a pair of states into a state of pair.

(|==|) :: Eq a => S a -> S a -> Bool

Checks equality of state by label, disregarding the flags.

labelMember :: Eq a => S a -> Set (S a) -> Bool

Checks if a state is member of a set of states by label, disregarding the flags.