Transition

Documentation

data T a

Transition of an automaton.

Constructors

T

Fields

transStart :: S a

The starting state of the transition.

symbol :: Char

The symbol of the transition.

transEnd :: S a

The ending state of the transition.

statesFromTrans :: Ord a => Set (T a) -> Set (S a)

Get all states from transitions.

transFromState :: Eq a => S a -> Set (T a) -> Set (T a)

Get all transitions which have this starting state.

transWithLabel :: Char -> Set (T a) -> Set (T a)

Get all transitions with this label.

endingStates :: Ord a => Set (T a) -> Set (S a)

Get all ending states of transitions.

createEndingState :: Ord a => Set (T a) -> S (Set a)

Create an ending state from transitions.

pairEndingStates :: Ord a => BoolRelation -> (T a, T b) -> S (a, b)

Pair ending states from a pair of transitions.

alphaFromTrans :: Set (T a) -> String

Get the alphabet from transitions.

alphaFromTransPair :: (T a, T b) -> String

Get the alphabet from a pair of transitions.

aFromTrans :: Set (T a) -> Char

Get the supposedly unique symbol from a set of transitions. Errors out if no such symbol exists.

aFromTransPair :: (T a, T b) -> Char

Get the supposedly unique symbol from a pair of transitions.

groupBySymbol :: Ord a => Set (T a) -> Set (Set (T a))

Groups a set of transitions by their symbol.